% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/other_functions.R
\name{ci_coverage}
\alias{ci_coverage}
\title{Calculating a Frequentist confidence interval covarage}
\usage{
ci_coverage(y_, y_hat_, sd_, prob_ = 0.5)
}
\description{
Calculating a Frequentist confidence interval covarage
}
