% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wrap_bart.R
\name{mvnbart}
\alias{mvnbart}
\title{Multivariate Normal Bayesian Additive Regression trees.}
\usage{
mvnbart(
  x_train,
  y_mat,
  x_test,
  n_tree = 100,
  node_min_size = 5,
  n_mcmc = 2000,
  n_burn = 500,
  alpha = 0.95,
  beta = 2,
  df = 3,
  sigquant = 0.9,
  kappa = 2,
  numcut = 100L,
  usequants = FALSE,
  m = 20,
  varimportance = TRUE,
  specify_variables = NULL
)
}
\arguments{
\item{x_train}{A \code{data.frame} of the training data covariates.}

\item{y_mat}{A numeric matrix of the data responses.}

\item{x_test}{A \code{data.frame} of the test data covariates.}

\item{n_tree}{The number of trees used in each set of trees for the respective \eqn{j} entry. The total number of trees is given by \eqn{j \times p}.}

\item{node_min_size}{The minimun number of observations within a terminal node.}

\item{n_mcmc}{The total number of MCMC iterations.}

\item{n_burn}{The number of MCMC iterations to be tretead as burn in samples}

\item{alpha}{The power parameter used in tree prior.}

\item{beta}{The base parameter used in tree prior.}

\item{df}{Degrees of freedom for the residuals variance prior. Not used when \eqn{Y} is a binary outcome.}

\item{sigquant}{Quantile used to define the residuals variance. Remind that the prior definition is based on \eqn{P(\sigma^{(j)} < \hat{\sigma}^{(j)})} where \eqn{\hat{\sigma}^{(j)}} is a "rough data-based estimation" for example, the sample variance of the observed \eqn{y^{(j)}} values.}

\item{kappa}{Hyper-parameter from the \eqn{\mu_{t\ell}^{(j)}} (Chipman et. al 2010). Such that \eqn{\sigma} = 0.25}

\item{numcut}{The maximum number of possible values used in the tree decision rules. The uniform approximation for choose a decision rule over \eqn{X^{(j)}} is given a grid of size \code{numcut}.}

\item{usequants}{Boolean; if true the quantiles are going to be used to define the grid of cutpoints.}

\item{m}{Hyperparameter used in the definition of the prior setting of the correlation matrix for the Probit-Multivariate approach.}

\item{varimportance}{Boolean; if true returns a matrix with \code{n_mcmc} rows and \eqn{d} columns corresponding to the total sum of number of times that a variable \eqn{j} was used among all trees of a MCMC iteration.}

\item{specify_variables}{a list of numeric vectors where each respective element contains the indexes of the covariates allowed to be selected for the set of trees for the respective respose \eqn{Y_{j}}. The default is \code{NULL} and allow to all covariates from \eqn{X} to be selected for all trees.}
}
\value{
In case of a continuous response the model returns
\code{mvnbart} object with the model predictions and parameters for the standard MVN. For the binary-classification problems the probit
Probit-MVN approach is used, and returns the \code{mvnbart} object.
}
\description{
\code{mvnbart()} function models a Bayesian Additive Regression trees model considering the Multivariate Normal (MVN) distribution
from the target dependent variable \eqn{Y_{i} \in \mathbb{R}^{d}}.
}
\details{
Add more details of the function description
}
