% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/other_functions.R
\name{pi_coverage}
\alias{pi_coverage}
\title{Coverage for the prediction intervals}
\usage{
pi_coverage(y, y_hat_post, sd_post, prob = 0.5, n_mcmc_replications = 1000)
}
\description{
Coverage for the prediction intervals
}
