% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulation_functions.R
\name{sim_class_mvn_friedman2}
\alias{sim_class_mvn_friedman2}
\title{Simulation setting for the Classification case}
\usage{
sim_class_mvn_friedman2(n, p, mvn_dim, Sigma = NULL)
}
\arguments{
\item{n}{sample size}

\item{p}{number of covariates}

\item{mvn_dim}{dimension \eqn{d} for the multivariate}

\item{Sigma}{Sigma function, default is \code{NULL}}
}
\value{
a simulated data
}
\description{
Simulation setting for the Classification case
}
