% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulation_functions.R
\name{sim_mvn_friedman1}
\alias{sim_mvn_friedman1}
\title{Simulation setting for the Regression case}
\usage{
sim_mvn_friedman1(n, p, mvn_dim, Sigma = NULL)
}
\arguments{
\item{n}{sample size}

\item{p}{number of covariates}

\item{mvn_dim}{dimension \eqn{d} for the multivariate}

\item{Sigma}{Sigma function, default is \code{NULL}}
}
\value{
a simulated data
}
\description{
Simulation setting for the Regression case
}
